move router to dedicated file
